requirements.txt
Flask
flask_sqlalchemy
requests
cassandra-driver
